add PYSECP_SO env variable parsing (if one do not want or cannot install to standard path); refactor
