+++ README.md
@@ -72,8 +72,13 @@ sudo make install
 if one builds secp256k1 without schnorrsig for example and then tries to import from it `from pysecp256k1.schnorrsig import schnorrsig_sign`
 `RuntimeError` is raised hinting that `libsecp256k1` is built without shnorrsig support. Same applies for all optional modules.
 
-This library uses the latest secp256k1 master. I only plan to release when secp256k1 releases. 
-So until then install like this 
+This library uses the latest secp256k1 master. If one needs to have older version of libsecp256k1 installed in 
+standard path and recent one for this library, do NOT do last step from above (`sudo make install`) and instead 
+export absolute path to desired `.so` file in environment variable.
+```shell
+export PYSECP_SO=/home/johndoe/secp256k1/.libs/libsecp256k1.so.0.0.0
+```
+I only plan to release when secp256k1 releases. So until then install like this: 
 ```shell
 python3 -m pip install -U pip wheel
 python3 -m pip install git+https://github.com/scgbckbone/python-secp256k1.git
